requirements.txt
fastapi==0.135.3
Flask==3.1.3
httpx==0.28.1
numpy==2.4.4
Pillow==12.2.0
psutil==7.0.0
pydantic==2.12.5
pytest==9.0.2
python-dotenv==1.2.2
python-multipart==0.0.20
Requests==2.33.1
torch==2.7.1
torchvision==0.22.1
uvicorn==0.43.0
